pyproject.toml
[tool.poetry]
name = "dop-zhu-zad"
version = "0.1.0"
description = ""
authors = ["dant4ick <lotuga.dan@gmail.com>"]
readme = "README.md"
packages = [{include = "dop_zhu_zad"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
passlib = "^1.7.4"
pydantic = "^2.7.4"
sqlalchemy = "^2.0.30"
python-jose = "^3.3.0"
pytest = "^8.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
